backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-docx==1.1.0
striprtf==0.0.27
langdetect==1.0.9
openai==1.12.0
httpx>=0.24.0,<0.28.0
